KeyResultTypes

+++ GoalMakerServer/GoalMakerServer/DTOS/KeyResultDTO.cs
@@ -8,7 +8,7 @@ namespace GoalMakerServer.DTOS
     public class KeyResultDTO
     {
         public string Name { get; set; }
-        public int PercentageOfSuccess { get; set; }
+        public double PercentageOfSuccess { get; set; }
         public int ConfidenceLevel { get; set; }
         public string Description { get; set; }
         public int Type { get; set; }
